Refactor toaster to use dedicated store, remove context usage, and reintroduce animations via constructable stylesheets
